Move changes in sprockets configurations to generation process start
