+++ decidim-generators/lib/decidim/generators/install_generator.rb
@@ -114,33 +114,6 @@ module Decidim
         rails "decidim_api:generate_docs"
       end
 
-      def remove_old_assets
-        remove_file "config/initializers/assets.rb"
-        remove_dir("app/assets")
-        remove_dir("app/javascript")
-      end
-
-      def remove_sprockets_requirement
-        gsub_file "config/application.rb", %r{require ['"]rails/all['"]\R}, <<~RUBY
-          require "decidim/rails"
-
-          # Add the frameworks used by your app that are not loaded by Decidim.
-          # require "action_mailbox/engine"
-          # require "action_text/engine"
-          require "action_cable/engine"
-          require "rails/test_unit/railtie"
-        RUBY
-
-        gsub_file "config/environments/development.rb", /config\.assets.*$/, ""
-        gsub_file "config/environments/test.rb", /config\.assets.*$/, ""
-        gsub_file "config/environments/production.rb", /config\.assets.*$/, ""
-      end
-
-      def add_sprockets_configuration
-        # Create an empty file so Sprockets 4 do not complain
-        create_file "app/assets/config/manifest.js"
-      end
-
       def letter_opener_web
         route <<~RUBY
           if Rails.env.development?
